requirements.txt
click==8.0.3
Flask==2.0.2
Flask-Cors==3.0.10
Flask-JWT-Extended==4.3.1
flask-swagger-ui==3.36.0
itsdangerous==2.0.1
Jinja2==3.0.3
MarkupSafe==2.0.1
psutil==5.9.0
PyJWT==2.3.0
python-dotenv==0.19.2
six==1.16.0
Werkzeug==2.0.2
